+++ CHANGELOG.md
@@ -6,6 +6,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [Checkpoint] - 2026-01-03 17:52:04
+
+### Changed Files
+- `src/features/expenses/ExpenseDashboardCompact.tsx`
+
+### Statistics
+```
+ 1 file changed, 14 insertions(+), 6 deletions(-)
+```
+
 ## [Checkpoint] - 2026-01-03 17:42:20
 
 ### Changed Files